requirements.txt
streamlit
mistralai
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
sentence-transformers
faiss-cpu
pypdf
tavily-python
beautifulsoup4
requests
python-dotenv